requirements.txt
pyyaml>=5.1.1
coolname>=1.1.0
tabulate>=0.8.3
tensorboardX>=1.4
runx
numpy
sklearn
h5py
jupyter
scikit-image
pillow
piexif
cffi
tqdm
dominate
opencv-python
nose
ninja